backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
PyJWT==2.9.0
firebase-admin==6.6.0
Werkzeug==3.0.4
python-dotenv==1.0.1
